Update server-sync structure
