+++ examples/server-sync/src/pages/Home.jsx
@@ -1,5 +1,5 @@
 import { useState } from "react";
-import { Button } from "@trussworks/react-uswds";
+import { Button, Alert, Link, GridContainer } from "@trussworks/react-uswds";
 import { Spinner, useOfflineStatus } from "@nmfs-radfish/react-radfish";
 import { MSW_ENDPOINT } from "../mocks/handlers";
 import { useOfflineStorage } from "@nmfs-radfish/react-radfish";
@@ -11,7 +11,7 @@ const dataIsSyncedMsg = "All data cached, ready to launch!";
 
 const HOME_BASE_DATA = "homebaseData";
 
-export const ServerSync = () => {
+export const HomePage = () => {
   const { isOffline } = useOfflineStatus();
   const { updateOfflineData, findOfflineData } = useOfflineStorage();
   const [isLoading, setIsLoading] = useState(false);
@@ -42,7 +42,7 @@ export const ServerSync = () => {
   const syncToHomebase = async () => {
     const lastSync = await findOfflineData(HOME_BASE_DATA);
     const lastSyncMsg = `Last sync at: ${findOfflineData("lastHomebaseSync")}`;
-    const { data: homebaseData } = await getRequestWithFetch(MSW_ENDPOINT.HOMEBASE); 
+    const { data: homebaseData } = await getRequestWithFetch(MSW_ENDPOINT.HOMEBASE);
     await updateOfflineData(HOME_BASE_DATA, homebaseData);
 
     await updateOfflineData("lastHomebaseSync", [{ uuid: "lastSynced", time: Date.now() }]);
@@ -79,17 +79,44 @@ export const ServerSync = () => {
   };
 
   return (
-    <div className="server-sync">
-      {isLoading ? (
-        <Spinner width={50} height={50} stroke={8} />
-      ) : (
-        <Button onClick={syncToHomebase}>Sync from Server</Button>
-      )}
-      <span
-        className={`${syncStatus.status ? "text-green" : "text-red"} margin-left-2 margin-top-2`}
-      >
-        {syncStatus.message + " " + lastSynced}
-      </span>
-    </div>
+    <>
+      <h1>Server Sync Example</h1>
+      <InfoAnnotation />
+      <div className="server-sync">
+        {isLoading ? (
+          <Spinner width={50} height={50} stroke={8} />
+        ) : (
+          <Button onClick={syncToHomebase}>Sync from Server</Button>
+        )}
+        <span
+          className={`${syncStatus.status ? "text-green" : "text-red"} margin-left-2 margin-top-2`}
+        >
+          {syncStatus.message + " " + lastSynced}
+        </span>
+      </div>
+    </>
   );
 };
+
+const InfoAnnotation = () => {
+  return (
+    <Alert type="info" heading="Information" headingLevel="h2">
+      This is an example that demonstrates a design pattern for syncing data from an API endpoint
+      into IndexedDB. The idea is that the application with fetch from the API, and store the
+      persistent data in IndexedDB for offline storage.
+      <br />
+      <br />
+      Please note that this is utilizing Mock Service Worker to intercept these API requests. In
+      production this would integrate with an external API.
+      <br />
+      <br />
+      The data is cached in index db and you can view it by opening the application tab in the
+      browser developer tools.
+      <Link href="https://nmfs-radfish.github.io/radfish" target="_blank" rel="noopener noreferrer">
+        <Button type="button">Go To Documentation</Button>
+      </Link>
+    </Alert>
+  );
+};
+
+export default HomePage;
