--- examples/server-sync/src/utilities/index.js
@@ -1,1 +1,0 @@
-export * from "./fieldValidators";
